change routes to support DIP4
