+++ packages/web/src/pages/auth/[property]/index.tsx
@@ -1,27 +1,26 @@
 import React from 'react'
+import { MarketOverview } from 'src/components/organisms/MarketOverview'
 import { Footer } from 'src/components/organisms/Footer'
 import { Header } from 'src/components/organisms/Header'
 import { Headline } from 'src/components/atoms/Headline'
 import { H2 } from 'src/components/atoms/Typography'
-import { PropertySelectForm } from 'src/components/organisms/PropertySelectForm'
 import { useRouter } from 'next/router'
 
 type Props = {}
 
-const AuthenticateNewAsset = (_: Props) => {
-  const { market } = useRouter().query as { market: string }
+const Auth = (_: Props) => {
+  const { property } = useRouter().query as { property: string }
   return (
     <>
       <Header />
       <Headline height={300}>
-        <H2>Authenticate a new asset</H2>
+        <H2>Choose Market</H2>
+        <div>Select one asset market to authenticate.</div>
       </Headline>
-      <div style={{ marginBottom: '78px' }}>
-        <PropertySelectForm market={market} />
-      </div>
+      <MarketOverview propertyAddress={property} />
       <Footer />
     </>
   )
 }
 
-export default AuthenticateNewAsset
+export default Auth
