+++ packages/web/src/pages/auth/index.tsx
@@ -1,24 +1,25 @@
 import React from 'react'
-import { MarketOverview } from 'src/components/organisms/MarketOverview'
 import { Footer } from 'src/components/organisms/Footer'
 import { Header } from 'src/components/organisms/Header'
 import { Headline } from 'src/components/atoms/Headline'
 import { H2 } from 'src/components/atoms/Typography'
+import { PropertySelectForm } from 'src/components/organisms/PropertySelectForm'
 
 type Props = {}
 
-const Auth = (_: Props) => {
+const AuthenticateNewAsset = (_: Props) => {
   return (
     <>
       <Header />
       <Headline height={300}>
-        <H2>Choose Market</H2>
-        <div>Select one asset market to authenticate.</div>
+        <H2>Authenticate a new asset</H2>
       </Headline>
-      <MarketOverview />
+      <div style={{ marginBottom: '78px' }}>
+        <PropertySelectForm />
+      </div>
       <Footer />
     </>
   )
 }
 
-export default Auth
+export default AuthenticateNewAsset
